Dev 1.12.0 (#100)
* ui engine upgrade to cef 107.1.12

* Migration to typescript and scss

* Pending: Its still not possible to delete files via updates

* New project for own updater

* Add missing phrases and words to autotranslation

* Remove font size upper limit

* Add server name to message sources for animated emotes

* Config window opens at 1/4 of primary display screen

* Change: config page layout

* Feature: gap between chat entries

* Feature: font size and menu size for high res monitors

* Feature: custom font size

* Feature: Assigned groups for chat messages can be updated

* Feature: color picker for channel specific sender

* Feature: Tabs can be filtered by groups

* Feature: auto scaling font size on config dialog for high res monitors

* Config: enable chatlog by default

+++ Gobchat.App/Core/Util/ArchiveUnpacker.cs
@@ -1,5 +1,5 @@
 ﻿/*******************************************************************************
- * Copyright (C) 2019-2022 MarbleBag
+ * Copyright (C) 2019-2023 MarbleBag
  *
  * This program is free software: you can redistribute it and/or modify it under
  * the terms of the GNU Affero General Public License as published by the Free
@@ -24,7 +24,7 @@ namespace Gobchat.Core.Util
 
         public bool DeleteArchiveOnCompletion { get; set; } = true;
 
-        public bool DeleteOutputFolderOnFail { get; set; } = false;
+        public bool DeleteOutputFolderOnFail { get; set; } = true;
 
         public ArchiveUnpacker(string archivePath, string outputFolderPath)
         {
@@ -47,7 +47,8 @@ namespace Gobchat.Core.Util
             catch (ExtractionFailedException ex)
             {
                 progressMonitor.Log(StringFormat.Format(Resources.GeneralErrorOccured, ex.Message));
-                DeleteExtractedData(progressMonitor);
+                if (DeleteOutputFolderOnFail)
+                    DeleteExtractedData(progressMonitor);
                 DeleteArchive(progressMonitor);
                 throw;
             }
@@ -61,7 +62,8 @@ namespace Gobchat.Core.Util
                     return Result.Completed;
 
                 case ArchiveUnpackerHelper.ExtractionResult.Canceled:
-                    DeleteExtractedData(progressMonitor);
+                    if (DeleteOutputFolderOnFail)
+                        DeleteExtractedData(progressMonitor);
                     return Result.Canceled;
             }
 
@@ -76,9 +78,8 @@ namespace Gobchat.Core.Util
 
         private void DeleteExtractedData(IProgressMonitor progressMonitor)
         {
-            progressMonitor.Log(Resources.Core_Util_ArchiveUnpacker_DeleteIncomplete);
-            if (DeleteOutputFolderOnFail)
-                Directory.Delete(OutputFolderPath, true);
+            progressMonitor.Log(Resources.Core_Util_ArchiveUnpacker_DeleteIncomplete);            
+            Directory.Delete(OutputFolderPath, true);
         }
 
         public enum Result
